Fixed some mappings
